Cache monthly metrics

+++ api/app.js
@@ -38,17 +38,6 @@ CACHE = config.redis && config.redis.enabled    ? true : false;
 
 if (process.argv.indexOf('debug')    !== -1) DEBUG = true;
 if (process.argv.indexOf('no-cache') !== -1) CACHE = false; 
-  
-if (CACHE) {
-  if (!config.redis || !config.redis.port || !config.redis.host) {
-    CACHE = false;
-    winston.error("Redis port and host are required");
-    
-  } else {
-    redis = require("redis").createClient(config.redis.port, config.redis.host, config.redis.options);
-  }
-}
-
 
 gatewayList = require('./gateways.json');
   // TODO find permanent location for gateways list
@@ -57,7 +46,6 @@ gatewayList = require('./gateways.json');
 DATEARRAY  = ['YYYY', '-MM', '-DD', 'THH', ':mm', ':ssZZ'];
 DATEFORMAT = DATEARRAY.join('');
   
-
 var apiRoutes = {
   'offers'                  : require("./routes/offers"),
   'offersexercised'         : require("./routes/offersExercised"),
@@ -145,15 +133,24 @@ function requestHandler(req, res) {
 //initialize ledger monitor
 monitor.ledgerMonitor();
 
-//do some cache intializations
+//cache initialization
 if (CACHE) {
+  if (!config.redis || !config.redis.port || !config.redis.host) {
+    CACHE = false;
+    winston.error("Redis port and host are required");
+    
+  } else {
+    redis = require("redis").createClient(config.redis.port, config.redis.host, config.redis.options);
    
-  //reset cache if the arg is present
-  if (process.argv.indexOf('reset-cache') !== -1) redis.flushdb(); 
+    //reset cache if the arg is present
+    if (process.argv.indexOf('reset-cache') !== -1) redis.flushdb(); 
   
-  redis.on("error", function (err) {
-    winston.error("Redis - " + err);
-    CACHE = false; //turn it off if its not working
-  });    
+    redis.on("error", function (err) {
+      winston.error("Redis - " + err);
+      CACHE = false; //turn it off if its not working
+    }); 
+    
+    //initialize historical metrics and associated cron jobs
+    //require('./library/history').init(); 
+  } 
 }
-
